# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DealPilot Local API
  version: 1.0.0
  description: |
    DealPilot 本地优先架构的后端 API 契约。
    所有端点前缀 /api/v1，除 /health 外全部需要 Bearer Token 认证。
    Agent 只监听 127.0.0.1，禁止 0.0.0.0。
    列表响应使用游标分页（cursor + next_cursor）。
    写操作支持 Idempotency-Key 请求头实现幂等重试。
  contact:
    name: DealPilot
servers:
  - url: http://127.0.0.1:{port}
    description: 本地 Agent 回环地址
...
paths:
...
  /api/v1/matches/resolve:
    post:
      tags: [matches]
      summary: 会话身份匹配
      description: 按平台和原始标识符标准化并查询客户，返回唯一/多候选/未命中。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MatchResolveRequest' }
      responses:
        '200':
          description: 匹配结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MatchResolveResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/matches/bind:
    post:
      tags: [matches]
      summary: 人工绑定会话到客户
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MatchBindRequest' }
      responses:
        '201':
          description: 绑定成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Binding' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [matches]
      summary: 解绑会话
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MatchUnbindRequest' }
      responses:
        '204':
          description: 解绑成功
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/follow-ups:
    get:
      tags: [follow-ups]
      summary: 跟进记录列表（按客户筛选，游标分页）
      parameters:
        - name: customer_id
          in: query
          schema: { type: string }
        - name: project_id
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 跟进列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FollowUpListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [follow-ups]
      summary: 新建跟进记录
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FollowUpCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FollowUp' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/follow-ups/{id}:
    put:
      tags: [follow-ups]
      summary: 编辑跟进记录
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FollowUpUpdate' }
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FollowUp' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      tags: [follow-ups]
      summary: 删除跟进记录
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      responses:
        '204':
          description: 删除成功
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/v1/projects/{id}/milestones:
    post:
      tags: [projects]
      summary: 添加里程碑
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MilestoneCreate' }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Milestone' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/milestones/{id}:
    put:
      tags: [projects]
      summary: 更新里程碑完成状态
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
              required: [completed]
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Milestone' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/imports/parse:
    post:
      tags: [imports]
      summary: 解析 Excel/CSV 并预览
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '200':
          description: 解析预览结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportParseResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/imports/{job_id}/commit:
    post:
      tags: [imports]
      summary: 确认提交导入
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportCommitRequest' }
      responses:
        '200':
          description: 导入结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportCommitResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/imports/{job_id}/errors:
    get:
      tags: [imports]
      summary: 下载错误报告
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: CSV 错误报告
          content:
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
    CursorParam:
      name: cursor
      in: query
      description: 游标分页位置
      schema: { type: string }
    LimitParam:
      name: limit
      in: query
      description: 每页数量
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      description: 幂等键，重复请求返回首次结果
      schema: { type: string }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 短期访问令牌，工作台来自 sessionStorage，插件来自 Native Messaging Bootstrap。
  responses:
    BadRequest:
      description: 请求参数或格式错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 未认证或令牌无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Origin 校验失败或权限不足
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: 资源冲突或状态非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: 错误码
        message:
          type: string
          description: 错误信息
        details:
          type: object
          description: 附加详情
        request_id:
          type: string
          description: 请求追踪 ID
      required: [code, message, request_id]
...
    CustomerGrade:
      type: string
      enum: [A, B, C]
    CustomerStatus:
      type: string
      enum: [active, inactive]
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string, nullable: true }
        country: { type: string, nullable: true }
        source: { type: string, nullable: true }
        grade: { $ref: '#/components/schemas/CustomerGrade' }
        status: { $ref: '#/components/schemas/CustomerStatus' }
        deleted_at: { type: string, nullable: true, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, name, grade, status, created_at, updated_at]
...
    Binding:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        platform: { type: string }
        raw_identifier: { type: string }
        normalized_identifier: { type: string }
        manually_bound: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, platform, raw_identifier, normalized_identifier, manually_bound, created_at]
    MatchResolveRequest:
      type: object
      properties:
        platform: { type: string }
        raw_identifier: { type: string }
      required: [platform, raw_identifier]
    MatchResolveResponse:
      type: object
      properties:
        status:
          type: string
          enum: [unique, multiple, none]
        candidates:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
        customer: { $ref: '#/components/schemas/Customer' }
      required: [status]
    MatchBindRequest:
      type: object
      properties:
        platform: { type: string }
        raw_identifier: { type: string }
        customer_id: { type: string }
      required: [platform, raw_identifier, customer_id]
    MatchUnbindRequest:
      type: object
      properties:
        platform: { type: string }
        raw_identifier: { type: string }
      required: [platform, raw_identifier]
    FollowUpType:
      type: string
      enum: [call, email, chat, visit, note, message]
    MessageDirection:
      type: string
      enum: [inbound, outbound]
    FollowUp:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        project_id: { type: string, nullable: true }
        type: { $ref: '#/components/schemas/FollowUpType' }
        note: { type: string, nullable: true }
        message_body: { type: string, nullable: true }
        message_direction: { $ref: '#/components/schemas/MessageDirection', nullable: true }
        occurred_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
      required: [id, customer_id, type, occurred_at, created_at]
    FollowUpCreate:
      type: object
      properties:
        customer_id: { type: string }
        project_id: { type: string }
        type: { $ref: '#/components/schemas/FollowUpType' }
        note: { type: string }
        message_body: { type: string }
        message_direction: { $ref: '#/components/schemas/MessageDirection' }
        occurred_at: { type: string, format: date-time }
      required: [customer_id, type, occurred_at]
    FollowUpUpdate:
      type: object
      properties:
        project_id: { type: string }
        type: { $ref: '#/components/schemas/FollowUpType' }
        note: { type: string }
        message_body: { type: string }
        message_direction: { $ref: '#/components/schemas/MessageDirection' }
        occurred_at: { type: string, format: date-time }
    FollowUpListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FollowUp' }
        next_cursor: { type: string, nullable: true }
      required: [items]
...
    Milestone:
      type: object
      properties:
        id: { type: string }
        project_id: { type: string }
        name: { type: string }
        date: { type: string, format: date }
        completed: { type: boolean }
        created_at: { type: string, format: date-time }
      required: [id, project_id, name, date, completed, created_at]
    MilestoneCreate:
      type: object
      properties:
        name: { type: string }
        date: { type: string, format: date }
      required: [name, date]
    ImportParseResponse:
      type: object
      properties:
        job_id: { type: string }
        total_rows: { type: integer }
        valid_rows: { type: integer }
        errors:
          type: array
          items:
            type: object
            properties:
              row: { type: integer }
              field: { type: string }
              message: { type: string }
        duplicate_candidates:
          type: array
          items:
            type: object
            properties:
              row: { type: integer }
              candidates: { type: array, items: { $ref: '#/components/schemas/Customer' } }
        preview:
          type: array
          items: { $ref: '#/components/schemas/Customer' }
      required: [job_id, total_rows, valid_rows]
    ImportCommitRequest:
      type: object
      properties:
        job_id: { type: string }
        resolutions:
          type: array
          items:
            type: object
            properties:
              row: { type: integer }
              action:
                type: string
                enum: [merge, skip, create]
              target_id: { type: string }
          description: 重复候选的处理决策
      required: [job_id]
    ImportCommitResponse:
      type: object
      properties:
        success: { type: integer }
        failed: { type: integer }
        skipped: { type: integer }
        duplicates: { type: integer }
      required: [success, failed, skipped, duplicates]
...
security:
  - BearerAuth: []